Fix all pyrefly type errors to reach 0 errors
Wraps nullable Discord attrs with not_none(), adds proper None guards and
type lifts, models heterogeneous config dicts with TypedDicts, and migrates
test session factories to async_sessionmaker. One-off scripts have a few
errors suppressed inline.

+++ services/draft_socket_client.py
@@ -17,7 +17,7 @@ class DraftSocketClient:
     def connected(self):
         return self.sio.connected
 
-    async def connect_with_retry(self, url: str, max_retries: int = 5, base_delay: int = 2):
+    async def connect_with_retry(self, url: str, max_retries: int = 5, base_delay: float = 2):
         """
         Attempts to connect to the given URL with exponential backoff.
         """
